perf: optimize AnnoyVectorIndex with async I/O and refactor factory
This commit replaces synchronous file I/O with asynchronous operations in `AnnoyVectorIndex` to prevent blocking the event loop. It also refactors the vector index factory for better maintainability and robustness.

Key changes:
- Converted `AnnoyVectorIndex.save` and `load` to use `fs/promises`.
- Renamed `src/lib/vector-index/index.js` to `factory.js`.
- Added explicit backend validation and prioritized options over env vars in the factory.
- Updated `SemanticCache` and `VectorIndex` interface to support the async persistence API.
- Fixed a CI failure by pinning `annoy.js` to `^2.1.6`.
- Fixed a bug in async testing in `factory.test.js`.
- Ensured all callers and tests are updated to `await` the async methods.
